Issue #3: Move toolbar related code to separate class.
