+++ app/src/main/java/org/mozilla/reference/browser/BrowserActivity.kt
@@ -15,6 +15,7 @@ import mozilla.components.concept.engine.EngineView
 import mozilla.components.concept.tabstray.TabsTray
 import mozilla.components.feature.intent.IntentProcessor
 import mozilla.components.support.utils.SafeIntent
+import org.mozilla.reference.browser.browser.BrowserFragment
 import org.mozilla.reference.browser.ext.components
 
 open class BrowserActivity : AppCompatActivity(), ComponentCallbacks2 {
