+++ gradle.properties
@@ -17,5 +17,5 @@ org.gradle.jvmargs=-Xmx1536m
 # org.gradle.parallel=true
 
 # We want to use AndroidX.
-android.useAndroidX=true
-android.enableJetifier=true
+# android.useAndroidX=true
+# android.enableJetifier=true
